[ENG-1060] Fix the `PlatformProvider` bug (#1353)
* Life is a lie and nothing makes sense

* consistent version

* Web uses the relative alias resolver it would seem

---------

+++ core/src/location/file_path_helper/mod.rs
@@ -156,7 +156,9 @@ pub fn path_is_hidden(path: &Path, metadata: &Metadata) -> bool {
 
 	#[cfg(target_family = "windows")]
 	{
-		const FILE_ATTRIBUTE_HIDDEN: u8 = 0x2;
+		use std::os::windows::fs::MetadataExt;
+
+		const FILE_ATTRIBUTE_HIDDEN: u32 = 0x2;
 
 		if (metadata.file_attributes() & FILE_ATTRIBUTE_HIDDEN) == FILE_ATTRIBUTE_HIDDEN {
 			return true;
